+++ app/Http/Controllers/BackEnd/Services/ServicesController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\BackEnd\Services;
 
+use App\Constant\PermissionConstant;
 use App\Http\Controllers\BackEnd\Categories\CategoriesController;
 use App\Http\Controllers\Controller;
 use App\Models\CategoriesServices;
@@ -10,6 +11,9 @@ use Cviebrock\EloquentSluggable\Services\SlugService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Lang;
 use App\Http\Requests\Backend\Services\ServicesFormRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+
 
 
 use function App\Helpers\uploadFile;
@@ -28,34 +32,44 @@ class ServicesController extends Controller
 
     public function index()
     {
-        $services = $this->services->paginate(5);
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_LIST)) {
+        $services = $this->services->sortable()->paginate(5);
         return view('backend.admin.services.index', compact('services'));
+        }
     }
 
     public function view(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_VIEW)) {
         $services = $this->services->find($request->id);
         $services->load('categories');
         return view('backend.admin.services.view', compact('services'));
+        }
     }
 
     public function delete(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_DELETE)) {
         $services = $this->services->find($request->id);
         $services->delete();
         return redirect()->route('backend.admin.services.show')->with('success', Lang::get('message.delete', ['model' => 'dịch vụ']));
+        }
     }
 
     public function servicesDelete(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_DELETE)) {
         $this->services->whereIn('id', explode(",", $request->ids))->delete();
         return response()->json(['success' => "Xóa dịch vụ thành công"]);
+        }
     }
 
     public function create()
     {
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_CREATE)) {
         $categories = $this->categories->all();
         return view('backend/admin/services/create', compact('categories'));
+        }
     }
 
     public function store(ServicesFormRequest $request)
@@ -70,16 +84,18 @@ class ServicesController extends Controller
             'description' => $request->description,
             'status' => $request->status,
             'time' => $request->time,
-            'slug' => SlugService::createSlug(Services::class, 'slug', $request->service_name),
+            'slug' => Str::slug($request->service_name),
         ]);
         return redirect()->route('backend.admin.services.show')->with('success', Lang::get('message.create', ['model' => 'dịch vụ']));
     }
 
     public function edit(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::SERVICES_PERMISSION_EDIT)) {
         $categories = $this->categories->all();
         $services = $this->services->find($request->id);
         return view('backend/admin/services/edit', compact('categories', 'services'));
+        }
     }
 
     public function update(Request $request)
@@ -101,7 +117,7 @@ class ServicesController extends Controller
             'description' => $request->description,
             'status' => $request->status,
             'time' => $request->time,
-            'slug' => SlugService::createSlug(Services::class, 'slug', $request->service_name),
+            'slug' => Str::slug($request->service_name),
         ]);
         return redirect()->route('backend.admin.services.show')->with('success', Lang::get('message.create', ['model' => 'dịch vụ']));
     }
